restructured sources
add coverage targets
