+++ sandbox/src/CMakeLists.txt
@@ -14,10 +14,9 @@ target_link_libraries(${PROJECT_NAME}
 	PRIVATE
 		soundEngine
 )
-
+target_code_coverage(${PROJECT_NAME})
 install(TARGETS ${PROJECT_NAME}
         RUNTIME DESTINATION bin
         LIBRARY DESTINATION lib
         ARCHIVE DESTINATION lib/${PROJECT_NAME})
 
-enable_coverage(${PROJECT_NAME})
