+++ soundengine/src/ALHelpers.h
