+++ soundengine/src/Buffer.h
