+++ soundengine/src/Loader.h
