+++ soundengine/src/Types.h
