--- src/CMakeLists.txt
@@ -1,3 +1,0 @@
-add_subdirectory(soundEngine)
-
-add_subdirectory(sandbox)
